week6/requirement.txt
Flask==3.0.0
python-dotenv==1.0.1
PyJWT==2.8.0